slotDetector/requirements.txt
selenium>=3.141.0
chromedriver-autoinstaller>=0.2.2
pyyaml==5.4.1